+++ lib/alerts_viewer_web/live/delay_alert_algorithm/median_and_standard_deviation_component_adherence.ex
@@ -81,7 +81,6 @@ defmodule AlertsViewer.DelayAlertAlgorithm.MedianAndStandardDeviationAdherenceCo
           <%= @min_std_val %>
         </span>
       </.controls_form>
-      <SnapshotButtonComponent.snapshot_button module_name={__MODULE__} />
     </div>
     """
   end
